refactor scraper module
